backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Configuração e env
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# NLP e ML
scikit-learn==1.5.2
nltk==3.9.1

# Upload/Parsing
python-multipart==0.0.9
pypdf==5.1.0

# (Opcional – se quiser integração com IA externas)
openai==1.56.0
google-generativeai==0.8.3
